requirements.txt
discord.py
pyfiglet
tinydb
aiofiles
tinydb-encrypted-jsonstorage